Restructuring pages and routes
